名前をわかりやすく修正
