+++ src/PipingServer.NET/Models/IPipingProvider.cs
@@ -5,19 +5,19 @@ using Microsoft.AspNetCore.Mvc;
 
 namespace Piping.Models
 {
-    public interface IWaiters : IDisposable
+    public interface IPipingProvider : IDisposable
     {
         public IActionResult AddReceiver(string RelativeUri, HttpContext Receiver) => AddReceiver(RelativeUri, Receiver.RequestAborted);
         IActionResult AddReceiver(string RelativeUri, CancellationToken Token = default);
         public IActionResult AddSender(string RelativeUri, HttpContext Context) => AddSender(RelativeUri, Context.Request, Context.RequestAborted);
         IActionResult AddSender(string RelativeUri, HttpRequest Request, CancellationToken Token = default);
     }
-    public interface IWaiter
+    public interface IPipe
     {
         RequestKey Key { get; }
-        WaiterStatus Status { get; }
+        PipeStatus Status { get; }
     }
-    public enum WaiterStatus
+    public enum PipeStatus
     {
         Wait,
         Ready,
